+++ src/Entities/Report/ReportType.php
@@ -20,7 +20,8 @@ use Paddle\SDK\PaddleEnum;
  * @method static ReportType ProductsPrices()
  * @method static ReportType Transactions()
  * @method static ReportType TransactionLineItems()
- * @method static ReportType Balance()
+ * @method static ReportType Balance() Deprecated: Use PayoutReconciliation() instead.
+ * @method static ReportType PayoutReconciliation()
  */
 final class ReportType extends PaddleEnum
 {
@@ -31,4 +32,5 @@ final class ReportType extends PaddleEnum
     private const Transactions = 'transactions';
     private const TransactionLineItems = 'transaction_line_items';
     private const Balance = 'balance';
+    private const PayoutReconciliation = 'payout_reconciliation';
 }